Increment the version in build.gradle to be 0.7.0-SNAPSHOT as agreed with Temi; Updated README.md file to reflect the latest instructions to set up a local development environment; Moved groovy files to their defined package (their compiled bytecode ends up in the defined package and not the current folder in the built war), to fix Errors messages detected by the Groovy Plug-in for Eclipse; Updated the defined packaged of ChartController.groovy to match the folder it is currently in ... there was no direct reference to this class and calling code in UrlMappings.groovy (in a different package) uses controller: 'chart' to refer to the ChartController, so changing the defined package is not expected to cause any issues; In build.gradle added a pathingJar to address https://github.com/gradle/gradle/issues/1989;
