+++ src/main/groovy/au/org/ala/soils2sat/LayerTreeNode.groovy
